requirements.txt
imutils==0.5.4
numpy==1.24.2
Pillow==9.4.0
opencv-python==4.7.0.72
tk==0.1.0